# src/cart/dto/update-cart-item.dto.ts
import { IsInt, IsNotEmpty, IsPositive } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class UpdateCartItemDto {
  @ApiProperty()
  @IsInt()
  @IsPositive()
  quantity: number;
}